+++ frontend/src/components/models/task/TaskColumn.tsx
@@ -15,7 +15,7 @@ import {
 import { TaskCreateModal } from 'components/models/task/TaskCreateModal'
 import { useInput } from 'hooks/useInput'
 import { usePopover } from 'hooks/usePopover'
-import { useControllTextArea } from 'hooks/useControlTextArea'
+import { useControllTextArea } from 'hooks/useControllTextArea'
 import { TaskList } from 'components/models/task/TaskList'
 import { CreateTaskButton } from 'components/ui/button/CreateTaskButton'
 import { SmallPopover } from 'components/ui/modal/SmallPopover'
